create common cli file modules
